#333: Moved the PassiveEsper*Parsers to the model folder
 They will be transformed into SkillEffects.
