+++ src/app/ffbe/mappers/effects/passive-effect-parser.factory.ts
@@ -1,6 +1,6 @@
 import {EffectParser} from './effect-parser';
 import {UnknownEffectParser} from './unknown-effect-parser';
-import {PassiveEsperDamageIncreaseParser} from './passives/passive-esper-damage-increase.parser';
+import {PassiveEsperDamageIncreaseParser} from '../../model/effects/passives/esper/passive-esper-damage-increase.parser';
 import {PassiveEquipmentCategoryUnlockParser} from './passives/passive-equipment-category-unlock.parser';
 import {PassiveStatsIncreaseHpThresholdParser} from './passives/passive-stats-increase-hp-threshold.parser';
 import {PassiveEquipmentCategoryStatsIncreaseParser} from './passives/passive-equipment-category-stats-increase.parser';
@@ -13,12 +13,12 @@ import {PassiveTargetChanceChangesParser} from './passives/passive-target-chance
 import {PassiveLbSpeedIncreaseParser} from './passives/passive-lb-speed-increase.parser';
 import {PassiveDualWieldWeaponCategoryUnlockParser} from './passives/passive-dual-wield-weapon-category-unlock.parser';
 import {PassiveDebuffsResistanceParser} from './passives/passive-debuffs-resistance.parser';
-import {PassiveEsperGroupSummonParser} from './passives/passive-esper-group-summon.parser';
+import {PassiveEsperGroupSummonParser} from '../../model/effects/passives/esper/passive-esper-group-summon.parser';
 import {PassiveJumpDamageIncreaseParser} from './passives/passive-jump-damage-increase.parser';
 import {PassiveSkillModifierIncreaseParser} from './passives/passive-skill-modifier-increase.parser';
 import {PassiveLbUpgradeParser} from './passives/passive-lb-upgrade.parser';
 import {PassiveLbUpgradeHpThresholdParser} from './passives/passive-lb-upgrade-hp-threshold.parser';
-import {PassiveEsperStatsIncreaseParser} from './passives/passive-esper-stats-increase.parser';
+import {PassiveEsperStatsIncreaseParser} from '../../model/effects/passives/esper/passive-esper-stats-increase.parser';
 import {PassiveCounterAttackChanceParser} from './passives/passive-counter-attack-chance.parser';
 import {PassiveMpDecreaseForSongsParser} from './passives/passive-mp-decrease-for-songs.parser';
 import {PassiveEvasionMagicalParser} from './passives/passive-evasion-magical.parser';
