+++ src/app/ffbe/model/effects/passives/esper/passive-esper-stats-increase.parser.ts
@@ -1,7 +1,7 @@
-import {EffectParser} from '../effect-parser';
-import {Skill} from '../../../model/skill.model';
-import {FFBE_ESPERS} from '../../../ffbe.constants';
-import {SkillEffect} from '../../../model/effects/skill-effect.model';
+import {EffectParser} from '../../../../mappers/effects/effect-parser';
+import {Skill} from '../../../skill.model';
+import {FFBE_ESPERS} from '../../../../ffbe.constants';
+import {SkillEffect} from '../../skill-effect.model';
 
 export class PassiveEsperStatsIncreaseParser extends EffectParser {
   public parse(effect: Array<any>, skill: Skill): string {
